+++ src/commands/visitAtPointCommands.ts
@@ -7,13 +7,13 @@ import { gitRun } from '../utils/gitRawRunner';
 import { CommitDetailView } from '../views/commitDetailView';
 import { views } from '../extension';
 import { StashItemView } from '../views/stashes/stashSectionView';
-import { StashDetailView } from '../views/stashDetailView';
 import { ChangeView } from '../views/changes/changeView';
 import { MagitCommit } from '../models/magitCommit';
 import { HunkView } from '../views/changes/hunkView';
 import { BranchListingView } from '../views/branches/branchListingView';
 import { RemoteBranchListingView } from '../views/remotes/remoteBranchListingView';
 import { TagListingView } from '../views/tags/tagListingView';
+import { showStashDetail } from './diffingCommands';
 
 export async function magitVisitAtPoint(repository: MagitRepository, currentView: DocumentView) {
 
@@ -44,17 +44,11 @@ export async function magitVisitAtPoint(repository: MagitRepository, currentView
   } else if (selectedView instanceof StashItemView) {
 
     const stash = (selectedView as StashItemView).stash;
-    const uri = StashDetailView.encodeLocation(repository, stash);
-
-    const result = await gitRun(repository, ['stash', 'show', '-p', `stash@{${stash.index}}`]);
-    const stashDiff = result.stdout;
-
-    views.set(uri.toString(), new StashDetailView(uri, stash, stashDiff));
-    workspace.openTextDocument(uri).then(doc => window.showTextDocument(doc, MagitUtils.oppositeActiveViewColumn()));
+    showStashDetail(repository, stash);
   }
 }
 
-async function visitCommit(repository: MagitRepository, commitHash: string) {
+export async function visitCommit(repository: MagitRepository, commitHash: string) {
   const commit: MagitCommit = { hash: commitHash, message: '', parents: [] };
   const result = await gitRun(repository, ['show', commitHash]);
   commit.diff = result.stdout;