+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ### 0.0.9
+- Adds diffing commands
 - Fixes issue with discarding remote branches
 - Visit ref heads commits from show-ref view